Move assert-display helpers into `style`
