--- src/assert-display.js
@@ -1,54 +1,0 @@
-/**
- * Copyright 2021 The AMP HTML Authors. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS-IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import {devError} from '#core/error';
-
-/**
- * Asserts that the style is not the `display` style.
- * This is the only possible way to pass a dynamic style to setStyle.
- *
- * If you wish to set `display`, use the `toggle` helper instead. This is so
- * changes to display can trigger necessary updates. See #17475.
- *
- * @param {string} style
- * @return {string}
- */
-export function assertNotDisplay(style) {
-  if (style === 'display') {
-    devError('STYLE', '`display` style detected. You must use toggle instead.');
-  }
-  return style;
-}
-
-/**
- * Asserts that the styles does not contain the `display` style.
- * This is the only possible way to pass a dynamic styles object to setStyles
- * and setImportantStyles.
- *
- * If you wish to set `display`, use the `toggle` helper instead. This is so
- * changes to display can trigger necessary updates. See #17475.
- *
- * @param {!Object<string, *>} styles
- * @return {!Object<string, *>}
- */
-export function assertDoesNotContainDisplay(styles) {
-  if ('display' in styles) {
-    devError(
-      'STYLE',
-      '`display` style detected in styles. You must use toggle instead.'
-    );
-  }
-  return styles;
-}
