+++ src/core/dom/style.js
@@ -16,6 +16,7 @@
 
 // Note: loaded by 3p system. Cannot rely on babel polyfills.
 import {devAssert} from '#core/assert';
+import {devError} from '#core/error';
 import {map} from '#core/types/object';
 
 /** @type {Object<string, string>} */
@@ -337,4 +338,42 @@ export function propagateObjectFitStyles(fromEl, toEl) {
  */
 function isVar(property) {
   return property.startsWith('--');
+}
+
+/**
+ * Asserts that the style is not the `display` style.
+ * This is the only possible way to pass a dynamic style to setStyle.
+ *
+ * If you wish to set `display`, use the `toggle` helper instead. This is so
+ * changes to display can trigger necessary updates. See #17475.
+ *
+ * @param {string} style
+ * @return {string}
+ */
+export function assertNotDisplay(style) {
+  if (style === 'display') {
+    devError('STYLE', '`display` style detected. You must use toggle instead.');
+  }
+  return style;
+}
+
+/**
+ * Asserts that the styles does not contain the `display` style.
+ * This is the only possible way to pass a dynamic styles object to setStyles
+ * and setImportantStyles.
+ *
+ * If you wish to set `display`, use the `toggle` helper instead. This is so
+ * changes to display can trigger necessary updates. See #17475.
+ *
+ * @param {!Object<string, *>} styles
+ * @return {!Object<string, *>}
+ */
+export function assertDoesNotContainDisplay(styles) {
+  if ('display' in styles) {
+    devError(
+      'STYLE',
+      '`display` style detected in styles. You must use toggle instead.'
+    );
+  }
+  return styles;
 }
